+++ docker/NEWS.rst
@@ -1,6 +1,15 @@
 Docker Container Release Notes
 ========================================================================
 
+Version 2.2.0:  October 5, 2018
+------------------------------------------------------------------------
+
+Pipeline Changes:
+
++ Fixed compress error messages in ``changeo-igblast`` and
+  ``changeo-clone``.
+
+
 Version 2.1.0:  September 20, 2018
 ------------------------------------------------------------------------